// src/app/shared/services/swagger/model/central.ts
/**
 * Demo API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Central { 
    centralId?: number;
    centralCode?: number;
    centralDescription?: string;
    centralCity?: string;
    centralRegion?: string;
    centralCountry?: string;
    centralParentId?: number;
}